engine2src phase 3 & 4
