+++ src/engine/canvas/index.ts
@@ -1,0 +1,2 @@
+export { default } from './canvas';
+export { default as Image } from './image';
