+++ src/engine/index.ts
@@ -10,13 +10,13 @@
  */
 
 export { Main } from './main';
-export { default as Canvas } from './canvas';
-export { default as Image } from './image';
-export { default as WebGL } from './webgl';
-export { default as Camera } from './camera';
-export { default as Ani } from './camera.ani';
-export { default as Kinetic } from './camera.kinetic';
-export { View, Coordinates, Viewport, DrawRect } from './shared';
-export { Mat4, Vec4 } from './webgl.mat';
-export { Bicubic, easeInOut, easeIn, easeOut, linear, mod1, modPI, twoNth, longitudeDistance } from './utils';
+export { default as TileCanvas } from './canvas/canvas';
+export { default as Image } from './canvas/image';
+export { default as SphericalView } from './webgl/webgl';
+export { default as Camera } from './camera/camera';
+export { default as Ani } from './camera/ani';
+export { default as Kinetic } from './camera/kinetic';
+export { View, Coordinates, Viewport, DrawRect } from './shared/shared';
+export { Mat4, Vec4 } from './webgl/mat';
+export { Bicubic, easeInOut, easeIn, easeOut, linear, mod1, modPI, twoNth, longitudeDistance } from './utils/utils';
 export { PI, PI2, PIh, segsX, segsY, base360Distance } from './globals';
