+++ src/engine/shared/index.ts
@@ -1,0 +1,1 @@
+export { View, Coordinates, Viewport, DrawRect } from './shared';
