+++ src/engine/webgl/index.ts
@@ -1,0 +1,2 @@
+export { default } from './webgl';
+export { Mat4, Vec4 } from './mat';
